listener/pyproject.toml
[tool.poetry]
name = "listener"
version = "0.1.0"
description = ""
authors = ["heiskane <henipe@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
pydantic = "^1.10.6"
paho-mqtt = "^1.6.1"
sqlalchemy = "^2.0.5.post1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
